+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* DataLocations
+  - added `addExtension()` method
+
 ## v0.1.0
 
 Released Thursday, 30th April 2020.